requirements.txt
absl-py>=0.13.0
astunparse>=1.6.3
Bottleneck>=1.3.2
cachetools>=4.2.2
certifi>=2023.7.22
charset-normalizer>=3.0.0
clang>=5.0
cycler>=0.10.0
flatbuffers>=1.12
fonttools>=4.25.0
gast>=0.4.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.5
google-pasta>=0.2.0
grpcio>=1.39.0
h5py>=3.6.0
idna>=3.4
joblib>=1.2.0
keras==2.6.0
Keras-Preprocessing>=1.1.2
kiwisolver>=1.3.2
Markdown>=3.3.4
matplotlib>=3.4.3
mkl-fft>=1.3.0
mkl-random>=1.2.2
mkl-service>=2.4.0
munkres>=1.1.4
numexpr>=2.8.3
numpy>=1.21.0
oauthlib>=3.1.1
olefile>=0.46
opencv-python>=4.5.3.56
opt-einsum>=3.3.0
pandas>=1.3.3
Pillow>=9.0.0
protobuf>=3.20.0
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
pyparsing>=2.4.7
python-dateutil>=2.8.2
pytz>=2023.3
requests>=2.31.0
requests-oauthlib>=1.3.0
rsa>=4.7.2
scikit-learn>=1.0.2
scipy>=1.7.1
seaborn>=0.11.2
sip>=4.19.13
six>=1.15.0
tensorboard==2.6.0
tensorboard-data-server>=0.6.1
tensorboard-plugin-wit>=1.8.0
tensorflow==2.16.1
tensorflow-estimator==2.6.0
termcolor>=1.1.0
threadpoolctl>=3.1.0
tornado>=6.1
typing-extensions>=3.10.0.0
urllib3>=1.26.6
Werkzeug>=2.0.1
wincertstore>=0.2
wrapt>=1.12.1

